backend/requirements.txt
anyio==4.8.0
asgiref==3.8.1
channels==4.2.0
channels_redis==4.2.1
click==8.1.8
Django==5.1.4
django-cors-headers==4.7.0
django-environ==0.11.2
django-graphql-jwt==0.4.0
djangorestframework==3.15.2
graphene==3.4.3
graphene-django==3.2.2
graphene-file-upload==1.3.0
graphql-core==3.2.6
graphql-relay==3.2.0
h11==0.14.0
httptools==0.6.4
idna==3.10
msgpack==1.1.0
pillow==11.1.0
promise==2.3
psycopg2==2.9.10
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
redis==5.2.1
servestatic==3.0.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
text-unidecode==1.3
typing_extensions==4.12.2
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.4
websockets==14.1
